requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1 # (DRF) для создания гибкого и мощного API для проекта
Pillow==9.0.0  # для загрузки картинок
psycopg2-binary==2.9.3 # для работы с базами данных
python-decouple==3.5 # для скрытия данных
pytz==2021.3
sqlparse==0.4.2

django-model-utils # использовала для выбора языка в фанфиках

drf-yasg # для подключения документации Swagger
